requirements.txt
# Language Detection App - Dependencies
# Install with: pip install -r requirements.txt

langdetect>=1.0.9
streamlit>=1.28.0
googletrans==4.0.0rc1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
